+++ data/models/src/events/builder/node.rs
@@ -2,6 +2,7 @@ use super::super::super::Node;
 use super::super::Event;
 use super::super::EventBuilder;
 use super::super::EventPayload;
+use super::super::NodeChanged;
 
 
 /// Build `Event`s that belongs to the node family.
@@ -15,7 +16,18 @@ impl NodeBuilder {
         NodeBuilder { builder }
     }
 
-    /// Build an `EventPayload::NodeNew`.
+    /// Build an `EventPayload::NodeChanged` event.
+    pub fn changed(self, before: Node, after: Node) -> Event {
+        let data = EventPayload::NodeChanged(NodeChanged {
+            cluster: before.cluster.clone(),
+            host: before.name.clone(),
+            before,
+            after,
+        });
+        self.builder.build(data)
+    }
+
+    /// Build an `EventPayload::NodeNew` event.
     pub fn node_new(self, node: Node) -> Event {
         let data = EventPayload::NodeNew(node);
         self.builder.build(data)
@@ -26,9 +38,26 @@ impl NodeBuilder {
 #[cfg(test)]
 mod tests {
     use replicante_agent_models::DatastoreInfo as WireNode;
-    use super::Node;
     use super::Event;
     use super::EventPayload;
+    use super::Node;
+    use super::NodeChanged;
+
+    #[test]
+    fn changed() {
+        let before = WireNode::new("cluster", "TestDB", "test", "1.2.3");
+        let before = Node::new(before);
+        let after = WireNode::new("cluster", "TestDB", "test", "4.5.6");
+        let after = Node::new(after);
+        let event = Event::builder().node().changed(before.clone(), after.clone());
+        let expected = EventPayload::NodeChanged(NodeChanged {
+            cluster: before.cluster.clone(),
+            host: before.name.clone(),
+            before,
+            after,
+        });
+        assert_eq!(event.payload, expected);
+    }
 
     #[test]
     fn node_new() {